miniamazon/docker-deploy/web-app/requirements.txt
Django>=2.0,<3.0
psycopg2
django-credit-cards
django-crispy-forms
